MyFunctionApp/requirements.txt
azure-functions
boto3
pymongo
minio
bson
azure.storage.blob
pytz
